chore(emit): split class_es5_ast_to_ir.rs below §19 cap (#10638)
Behavior-neutral relocation of cohesive method groups from the central
`AstToIr` impl block into `#[path]` submodules, mirroring the existing
`async_es5_ir_*.rs` pattern. The main file goes from 2985 LOC → 1869 LOC,
under §19's hard 2000-line cap.

Extracted submodules:
- `class_es5_ast_to_ir_comments.rs` (289 LOC): trailing-comment attachment
  policy and the three associated unit tests.
- `class_es5_ast_to_ir_control_flow.rs` (260 LOC): `throw`, `try`, loops,
  `switch`, `break`/`continue`, labeled statements.
- `class_es5_ast_to_ir_for_in_of.rs` (158 LOC): `for-in`/`for-of` with the
  ES5 array-indexing fast path.
- `class_es5_ast_to_ir_expressions.rs` (473 LOC): calls (including dynamic
  `import()` and `super.x()`), property/element access, binary/unary
  operators, parenthesized, and conditional expressions.

Cross-module-called methods are `pub(super) fn`; helpers used only inside
the same submodule stay `fn`. `emit_ir_fragment_to_string` is elevated to
`pub(super)` because `convert_await_expression` (still in the main file)
calls it.

Verification: `cargo check -p tsz-emitter` and
`cargo clippy -p tsz-emitter --all-targets -- -D warnings` are clean;
`cargo test -p tsz-emitter --lib` shows the same 3102 passed / 35 failed /
1 ignored counts as `origin/main` (the 35 failures are pre-existing).

The 4918-line `async_es5_ir.rs` stays as a follow-up so this first slice
is reviewable in isolation.

AgentName: opus47-1m-vibrant-lovelace-zFExS
