+++ crates/tsz-emitter/src/transforms/class_es5_ast_to_ir.rs
@@ -8,11 +8,20 @@ use crate::context::transform::TransformDirective;
 use crate::transforms::async_es5_ir::AsyncES5Transformer;
 use rustc_hash::FxHashSet;
 use tsz_common::common::ModuleKind;
-use tsz_parser::parser::node::{ForInOfData, FunctionData, Node};
+use tsz_parser::parser::node::{FunctionData, Node};
 use tsz_parser::parser::node_flags;
 use tsz_parser::syntax::transform_utils::{
     contains_new_target_reference, contains_super_reference,
 };
+
+#[path = "class_es5_ast_to_ir_comments.rs"]
+mod comments;
+#[path = "class_es5_ast_to_ir_control_flow.rs"]
+mod control_flow;
+#[path = "class_es5_ast_to_ir_expressions.rs"]
+mod expressions;
+#[path = "class_es5_ast_to_ir_for_in_of.rs"]
+mod for_in_of;
 
 #[derive(Clone)]
 enum ThisSubstitution {
@@ -419,234 +428,6 @@ impl<'a> AstToIr<'a> {
         };
 
         self.attach_trailing_comment(node, statement)
-    }
-
-    fn attach_trailing_comment(
-        &self,
-        node: &tsz_parser::parser::node::Node,
-        statement: IRNode,
-    ) -> IRNode {
-        let Some(source_text) = self.source_text else {
-            return statement;
-        };
-        if let Some(comment_text) = self.included_trailing_line_comment(node, source_text) {
-            return IRNode::Sequence(vec![
-                statement,
-                IRNode::TrailingComment(comment_text.into()),
-            ]);
-        }
-        let scan_start =
-            Self::find_actual_statement_end(source_text, node.pos as usize, node.end as usize);
-        for comment in crate::emitter::get_trailing_comment_ranges(source_text, scan_start) {
-            if !self.comment_starts_before_limit(comment.pos) {
-                continue;
-            }
-            let gap_start = scan_start.min(source_text.len());
-            let gap_end = (comment.pos as usize).min(source_text.len());
-            if gap_start > gap_end
-                || !Self::can_attach_trailing_comment_gap(&source_text[gap_start..gap_end])
-            {
-                continue;
-            }
-            let comment_text = &source_text[comment.pos as usize..comment.end as usize];
-            let trimmed = comment_text.trim_start();
-            if trimmed.starts_with("//") || trimmed.starts_with("/*") {
-                return IRNode::Sequence(vec![
-                    statement,
-                    IRNode::TrailingComment(comment_text.to_string().into()),
-                ]);
-            }
-        }
-        if node.kind == syntax_kind_ext::RETURN_STATEMENT {
-            return statement;
-        }
-        let line_start = scan_start.min(source_text.len());
-        let line_end = source_text[line_start..]
-            .find(['\n', '\r'])
-            .map_or(source_text.len(), |offset| line_start + offset);
-        let line = &source_text[line_start..line_end];
-        if let Some(comment_start) = Self::attached_line_comment_start(node.kind, line) {
-            let comment_pos = line_start + comment_start;
-            if !self.comment_starts_before_limit(comment_pos as u32) {
-                return statement;
-            }
-            let comment_text = &line[comment_start..];
-            return IRNode::Sequence(vec![
-                statement,
-                IRNode::TrailingComment(comment_text.to_string().into()),
-            ]);
-        }
-        statement
-    }
-
-    /// Scan `[node_start, node_end)` and return the position after the last depth-0
-    /// statement terminator (`;` or closing bracket). The parser sets `node.end` to
-    /// the NEXT token's end, so we need to scan back to the actual statement boundary
-    /// before searching for trailing comments.
-    fn find_actual_statement_end(source_text: &str, node_start: usize, node_end: usize) -> usize {
-        let bytes = source_text.as_bytes();
-        let end = node_end.min(bytes.len());
-        let start = node_start.min(end);
-
-        let mut depth: i32 = 0;
-        let mut last_stmt_end = end;
-        let mut i = start;
-        let mut in_string: Option<u8> = None;
-
-        while i < end {
-            let ch = bytes[i];
-            if let Some(quote) = in_string {
-                if ch == b'\\' {
-                    i = (i + 2).min(end);
-                    continue;
-                }
-                if ch == quote {
-                    in_string = None;
-                }
-                i += 1;
-                continue;
-            }
-            match ch {
-                b'\'' | b'"' | b'`' => {
-                    in_string = Some(ch);
-                    i += 1;
-                }
-                b'/' if i + 1 < end && bytes[i + 1] == b'/' => {
-                    i += 2;
-                    while i < end && bytes[i] != b'\n' && bytes[i] != b'\r' {
-                        i += 1;
-                    }
-                }
-                b'/' if i + 1 < end && bytes[i + 1] == b'*' => {
-                    i += 2;
-                    while i + 1 < end && !(bytes[i] == b'*' && bytes[i + 1] == b'/') {
-                        i += 1;
-                    }
-                    if i + 1 < end {
-                        i += 2;
-                    }
-                }
-                b'{' | b'(' | b'[' => {
-                    depth += 1;
-                    i += 1;
-                }
-                b'}' | b')' | b']' => {
-                    depth = (depth - 1).max(0);
-                    if depth == 0 {
-                        last_stmt_end = i + 1;
-                    }
-                    i += 1;
-                }
-                b';' if depth == 0 => {
-                    last_stmt_end = i + 1;
-                    i += 1;
-                }
-                _ => {
-                    i += 1;
-                }
-            }
-        }
-
-        last_stmt_end
-    }
-
-    fn included_trailing_line_comment(
-        &self,
-        node: &tsz_parser::parser::node::Node,
-        source_text: &str,
-    ) -> Option<String> {
-        let node_start = (node.pos as usize).min(source_text.len());
-        let node_end = self
-            .trailing_comment_limit
-            .get()
-            .map_or(node.end as usize, |limit| {
-                (limit as usize).min(node.end as usize)
-            })
-            .min(source_text.len());
-        if node_start >= node_end {
-            return None;
-        }
-
-        let mut scan_end = node_end;
-        while scan_end > node_start {
-            let Some(ch) = source_text[..scan_end].chars().next_back() else {
-                break;
-            };
-            if !ch.is_whitespace() {
-                break;
-            }
-            scan_end -= ch.len_utf8();
-        }
-        if node_start >= scan_end {
-            return None;
-        }
-
-        let line_start = source_text[..scan_end]
-            .rfind(['\n', '\r'])
-            .map_or(node_start, |offset| (offset + 1).max(node_start));
-        let line = &source_text[line_start..scan_end];
-        let comment_start = Self::line_comment_start(line)?;
-        let comment_pos = line_start + comment_start;
-        if comment_pos < node_start || !self.comment_starts_before_limit(comment_pos as u32) {
-            return None;
-        }
-        Some(line[comment_start..].trim_end().to_string())
-    }
-
-    fn comment_starts_before_limit(&self, comment_pos: u32) -> bool {
-        self.trailing_comment_limit
-            .get()
-            .is_none_or(|limit| comment_pos < limit)
-    }
-
-    fn attached_line_comment_start(node_kind: u16, line_suffix: &str) -> Option<usize> {
-        let comment_start = Self::line_comment_start(line_suffix)?;
-        let gap = &line_suffix[..comment_start];
-        let gap_belongs_to_statement = if node_kind == syntax_kind_ext::EXPRESSION_STATEMENT {
-            Self::expression_statement_trailing_comment_gap(gap)
-        } else {
-            Self::can_attach_trailing_comment_gap(gap)
-        };
-        gap_belongs_to_statement.then_some(comment_start)
-    }
-
-    fn expression_statement_trailing_comment_gap(gap: &str) -> bool {
-        gap.bytes()
-            .all(|byte| matches!(byte, b' ' | b'\t' | b';' | b')' | b']' | b'}'))
-    }
-
-    fn can_attach_trailing_comment_gap(gap: &str) -> bool {
-        gap.chars().all(|ch| ch.is_whitespace() || ch == ';')
-    }
-
-    fn line_comment_start(line: &str) -> Option<usize> {
-        let bytes = line.as_bytes();
-        let mut i = 0;
-        let mut quote = None;
-        while i + 1 < bytes.len() {
-            let byte = bytes[i];
-            if let Some(delim) = quote {
-                if byte == b'\\' {
-                    i += 2;
-                    continue;
-                }
-                if byte == delim {
-                    quote = None;
-                }
-                i += 1;
-                continue;
-            }
-            if byte == b'\'' || byte == b'"' || byte == b'`' {
-                quote = Some(byte);
-                i += 1;
-                continue;
-            }
-            if byte == b'/' && bytes[i + 1] == b'/' {
-                return Some(i);
-            }
-            i += 1;
-        }
-        None
     }
 
     /// Convert an expression to IR
@@ -1142,402 +923,6 @@ impl<'a> AstToIr<'a> {
         }
     }
 
-    fn convert_throw_statement(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        // Throw uses ReturnData (same as return statement)
-        if let Some(return_data) = self.arena.get_return_statement(node) {
-            IRNode::ThrowStatement(Box::new(self.convert_expression(return_data.expression)))
-        } else {
-            IRNode::ASTRef(idx)
-        }
-    }
-
-    fn convert_try_statement(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        if let Some(try_data) = self.arena.get_try(node) {
-            let try_block = Box::new(self.convert_statement(try_data.try_block));
-
-            let catch_clause = if try_data.catch_clause.is_none() {
-                None
-            } else if let Some(catch_node) = self.arena.get(try_data.catch_clause)
-                && let Some(catch) = self.arena.get_catch_clause(catch_node)
-            {
-                let param = if catch.variable_declaration.is_none() {
-                    None
-                } else {
-                    get_identifier_text(self.arena, catch.variable_declaration)
-                };
-                let catch_block = self.arena.get(catch.block);
-                let body = if let Some(block_node) = catch_block
-                    && let Some(block) = self.arena.get_block(block_node)
-                {
-                    block
-                        .statements
-                        .nodes
-                        .iter()
-                        .map(|&s| self.convert_statement(s))
-                        .collect()
-                } else {
-                    vec![]
-                };
-                Some(IRCatchClause {
-                    param: param.map(Into::into),
-                    body,
-                    single_line: false,
-                })
-            } else {
-                None
-            };
-
-            let finally_block = if try_data.finally_block.is_none() {
-                None
-            } else {
-                Some(Box::new(self.convert_statement(try_data.finally_block)))
-            };
-
-            IRNode::TryStatement {
-                try_block,
-                catch_clause,
-                finally_block,
-            }
-        } else {
-            IRNode::ASTRef(idx)
-        }
-    }
-
-    fn convert_for_statement(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        // For uses LoopData (same as while/do-while)
-        if let Some(loop_data) = self.arena.get_loop(node) {
-            let initializer = if loop_data.initializer.is_none() {
-                None
-            } else {
-                Some(Box::new(self.convert_expression(loop_data.initializer)))
-            };
-            let condition = if loop_data.condition.is_none() {
-                None
-            } else {
-                Some(Box::new(self.convert_expression(loop_data.condition)))
-            };
-            let incrementor = if loop_data.incrementor.is_none() {
-                None
-            } else {
-                Some(Box::new(self.convert_expression(loop_data.incrementor)))
-            };
-            IRNode::ForStatement {
-                initializer,
-                condition,
-                incrementor,
-                body: Box::new(self.convert_statement(loop_data.statement)),
-            }
-        } else {
-            IRNode::ASTRef(idx)
-        }
-    }
-
-    fn convert_while_statement(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        // While uses LoopData (same as for/do-while)
-        if let Some(loop_data) = self.arena.get_loop(node) {
-            IRNode::WhileStatement {
-                condition: Box::new(self.convert_expression(loop_data.condition)),
-                body: Box::new(self.convert_statement(loop_data.statement)),
-            }
-        } else {
-            IRNode::ASTRef(idx)
-        }
-    }
-
-    fn convert_do_while_statement(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        // DoWhile uses LoopData (same as while/for loops)
-        if let Some(loop_data) = self.arena.get_loop(node) {
-            IRNode::DoWhileStatement {
-                body: Box::new(self.convert_statement(loop_data.statement)),
-                condition: Box::new(self.convert_expression(loop_data.condition)),
-            }
-        } else {
-            IRNode::ASTRef(idx)
-        }
-    }
-
-    fn convert_switch_statement(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        if let Some(switch_data) = self.arena.get_switch(node) {
-            // Case block uses BlockData where statements contains the case clauses
-            let cases = if let Some(case_block_node) = self.arena.get(switch_data.case_block)
-                && let Some(block_data) = self.arena.get_block(case_block_node)
-            {
-                block_data
-                    .statements
-                    .nodes
-                    .iter()
-                    .map(|&c| self.convert_switch_case(c))
-                    .collect()
-            } else {
-                vec![]
-            };
-            IRNode::SwitchStatement {
-                expression: Box::new(self.convert_expression(switch_data.expression)),
-                cases,
-            }
-        } else {
-            IRNode::ASTRef(idx)
-        }
-    }
-
-    fn convert_switch_case(&self, idx: NodeIndex) -> IRSwitchCase {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        // get_case_clause works for both CASE_CLAUSE and DEFAULT_CLAUSE
-        // For DEFAULT_CLAUSE, expression is NONE
-        if let Some(case_clause) = self.arena.get_case_clause(node) {
-            let test = if case_clause.expression.is_none() {
-                None // Default clause
-            } else {
-                Some(self.convert_expression(case_clause.expression))
-            };
-            IRSwitchCase {
-                test,
-                statements: case_clause
-                    .statements
-                    .nodes
-                    .iter()
-                    .map(|&s| self.convert_statement(s))
-                    .collect(),
-                inline: false,
-            }
-        } else {
-            IRSwitchCase {
-                test: None,
-                statements: vec![],
-                inline: false,
-            }
-        }
-    }
-
-    fn convert_break_statement(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        if let Some(jump_data) = self.arena.get_jump_data(node) {
-            let label = if jump_data.label.is_none() {
-                None
-            } else {
-                get_identifier_text(self.arena, jump_data.label)
-            };
-            IRNode::BreakStatement(label.map(Into::into))
-        } else {
-            IRNode::BreakStatement(None)
-        }
-    }
-
-    fn convert_continue_statement(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        if let Some(jump_data) = self.arena.get_jump_data(node) {
-            let label = if jump_data.label.is_none() {
-                None
-            } else {
-                get_identifier_text(self.arena, jump_data.label)
-            };
-            IRNode::ContinueStatement(label.map(Into::into))
-        } else {
-            IRNode::ContinueStatement(None)
-        }
-    }
-
-    fn convert_labeled_statement(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        if let Some(labeled) = self.arena.get_labeled_statement(node)
-            && let Some(label) = get_identifier_text(self.arena, labeled.label)
-        {
-            if self.emit_await_as_yield && label == "await" {
-                return IRNode::Sequence(vec![
-                    IRNode::expr_stmt(IRNode::Raw("yield ".into())),
-                    self.convert_statement(labeled.statement),
-                ]);
-            }
-            return IRNode::LabeledStatement {
-                label: label.into(),
-                statement: Box::new(self.convert_statement(labeled.statement)),
-            };
-        }
-        IRNode::ASTRef(idx)
-    }
-
-    fn convert_for_in_of_statement(&self, idx: NodeIndex) -> IRNode {
-        let Some(node) = self.arena.get(idx) else {
-            return IRNode::ASTRef(idx);
-        };
-        if node.kind == syntax_kind_ext::FOR_OF_STATEMENT
-            && self.can_delegate_es5_for_of_to_ast_printer(idx)
-        {
-            return IRNode::ASTRef(idx);
-        }
-
-        // Issue #3539: previously we returned `ASTRef(idx)` which delegates
-        // to the AST printer that has no `_this` substitution context. The
-        // body of `for-in`/`for-of` inside a derived ES5 constructor must
-        // recurse through `convert_statement` so any `this` reference
-        // becomes `_this`.
-        let Some(loop_data) = self.arena.get_for_in_of(node) else {
-            return IRNode::ASTRef(idx);
-        };
-        if node.kind == syntax_kind_ext::FOR_OF_STATEMENT
-            && !self.downlevel_iteration
-            && !loop_data.await_modifier
-            && let Some(ir) = self.convert_for_of_array_indexing(loop_data)
-        {
-            return ir;
-        }
-        let kind = if node.kind == syntax_kind_ext::FOR_OF_STATEMENT {
-            if loop_data.await_modifier {
-                std::borrow::Cow::Borrowed("await of")
-            } else {
-                std::borrow::Cow::Borrowed("of")
-            }
-        } else {
-            std::borrow::Cow::Borrowed("in")
-        };
-        // The initializer is either a VariableDeclarationList or an
-        // expression. The variable-declaration case never references
-        // `this`, so we keep it as ASTRef for simplicity. The expression
-        // case still benefits from substitution, so use convert_expression.
-        // The initializer is either a VariableDeclarationList or an
-        // expression. For the variable-declaration case we synthesize
-        // `var <name>` from the parsed VariableData rather than slicing
-        // the source — the parser includes the trailing `of`/`in` keyword
-        // in the VARIABLE_DECLARATION_LIST node's range, so an `ASTRef`
-        // would re-emit the keyword. The expression case still benefits
-        // from substitution, so use convert_expression.
-        let initializer_node = self.arena.get(loop_data.initializer);
-        let initializer = if let Some(init_node) = initializer_node
-            && init_node.kind == syntax_kind_ext::VARIABLE_DECLARATION_LIST
-            && let Some(var_data) = self.arena.get_variable(init_node)
-        {
-            let mut text = String::from("var ");
-            for (i, &decl_idx) in var_data.declarations.nodes.iter().enumerate() {
-                if i > 0 {
-                    text.push_str(", ");
-                }
-                if let Some(decl_node) = self.arena.get(decl_idx)
-                    && let Some(decl) = self.arena.get_variable_declaration(decl_node)
-                {
-                    text.push_str(&crate::transforms::emit_utils::identifier_text_or_empty(
-                        self.arena, decl.name,
-                    ));
-                }
-            }
-            IRNode::Raw(text.into())
-        } else {
-            self.convert_expression(loop_data.initializer)
-        };
-        IRNode::ForInOfStatement {
-            kind,
-            initializer: Box::new(initializer),
-            expression: Box::new(self.convert_expression(loop_data.expression)),
-            body: Box::new(self.convert_statement(loop_data.statement)),
-            multiline_body: false,
-        }
-    }
-
-    fn convert_for_of_array_indexing(&self, loop_data: &ForInOfData) -> Option<IRNode> {
-        let init_node = self.arena.get(loop_data.initializer)?;
-        if init_node.kind != syntax_kind_ext::VARIABLE_DECLARATION_LIST {
-            return None;
-        }
-        let var_data = self.arena.get_variable(init_node)?;
-        if var_data.declarations.nodes.len() != 1 {
-            return None;
-        }
-        let decl_idx = var_data.declarations.nodes[0];
-        let decl_node = self.arena.get(decl_idx)?;
-        let decl = self.arena.get_variable_declaration(decl_node)?;
-        let binding_name = get_identifier_text(self.arena, decl.name)?;
-
-        let index_name = if self.temp_var_counter.get() == 0 && !self.source_has_identifier("_i") {
-            "_i".to_string()
-        } else {
-            self.generate_temp_name()
-        };
-        let array_name = self.generate_temp_name();
-        let iterable = self.convert_expression(loop_data.expression);
-
-        let mut body = vec![IRNode::VarDecl {
-            name: binding_name.into(),
-            initializer: Some(Box::new(IRNode::elem(
-                IRNode::id(array_name.clone()),
-                IRNode::id(index_name.clone()),
-            ))),
-        }];
-        body.extend(self.convert_loop_body_statements(loop_data.statement));
-
-        Some(IRNode::ForStatement {
-            initializer: Some(Box::new(IRNode::VarDeclList(vec![
-                IRNode::VarDecl {
-                    name: index_name.clone().into(),
-                    initializer: Some(Box::new(IRNode::number("0"))),
-                },
-                IRNode::VarDecl {
-                    name: array_name.clone().into(),
-                    initializer: Some(Box::new(iterable)),
-                },
-            ]))),
-            condition: Some(Box::new(IRNode::BinaryExpr {
-                left: Box::new(IRNode::id(index_name.clone())),
-                operator: "<".into(),
-                right: Box::new(IRNode::prop(IRNode::id(array_name), "length")),
-            })),
-            incrementor: Some(Box::new(IRNode::PostfixUnaryExpr {
-                operand: Box::new(IRNode::id(index_name)),
-                operator: "++".into(),
-            })),
-            body: Box::new(IRNode::Block(body)),
-        })
-    }
-
-    fn convert_loop_body_statements(&self, statement_idx: NodeIndex) -> Vec<IRNode> {
-        if let Some(statement_node) = self.arena.get(statement_idx)
-            && statement_node.kind == syntax_kind_ext::BLOCK
-            && let Some(block) = self.arena.get_block(statement_node)
-        {
-            return block
-                .statements
-                .nodes
-                .iter()
-                .map(|&stmt_idx| self.convert_statement(stmt_idx))
-                .collect();
-        }
-        vec![self.convert_statement(statement_idx)]
-    }
-
     fn convert_identifier(&self, idx: NodeIndex) -> IRNode {
         let node = self
             .arena
@@ -1578,464 +963,6 @@ impl<'a> AstToIr<'a> {
     const fn convert_string_literal(&self, idx: NodeIndex) -> IRNode {
         // Use ASTRef to preserve original quote style from source text
         IRNode::ASTRef(idx)
-    }
-
-    fn convert_call_expression(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        if let Some(call) = self.arena.get_call_expr(node) {
-            let args: Vec<IRNode> = if let Some(ref args) = call.arguments {
-                args.nodes
-                    .iter()
-                    .map(|&a| self.convert_expression(a))
-                    .collect()
-            } else {
-                vec![]
-            };
-
-            if matches!(
-                self.module_kind,
-                ModuleKind::AMD | ModuleKind::UMD | ModuleKind::System
-            ) && let Some(callee_node) = self.arena.get(call.expression)
-                && callee_node.kind == SyntaxKind::ImportKeyword as u16
-            {
-                return self.convert_wrapped_dynamic_import(call.arguments.as_ref());
-            }
-
-            // Check for bare super(args) → _this = _super.call(this, args) || this
-            // This handles super() in expression contexts (e.g. computed property names).
-            if self.has_super
-                && let Some(cn) = self.arena.get(call.expression)
-                && cn.kind == SyntaxKind::SuperKeyword as u16
-            {
-                let mut call_args = vec![IRNode::this()];
-                call_args.extend(args);
-                // _this = _super.call(this, args...) || this
-                return IRNode::assign(
-                    IRNode::id("_this"),
-                    IRNode::logical_or(
-                        IRNode::call(
-                            IRNode::prop(IRNode::id(self.super_name.clone()), "call"),
-                            call_args,
-                        ),
-                        IRNode::this(),
-                    ),
-                );
-            }
-
-            // Check for super.method(args) or super[expr](args) → _super.prototype.method.call(this, args)
-            if let Some(super_call) = self.try_convert_super_method_call(
-                call.expression,
-                args.clone(),
-                node.is_optional_chain(),
-            ) {
-                return super_call;
-            }
-
-            let callee = self.convert_expression(call.expression);
-            IRNode::CallExpr {
-                callee: Box::new(callee),
-                arguments: args,
-            }
-        } else {
-            IRNode::ASTRef(idx)
-        }
-    }
-
-    fn convert_wrapped_dynamic_import(&self, args: Option<&NodeList>) -> IRNode {
-        let first_arg = self.first_dynamic_import_argument(args);
-        let first_arg_is_string_like = first_arg.is_none_or(|arg| {
-            crate::transforms::emit_utils::dynamic_import_arg_is_string_like(self.arena, arg)
-        });
-
-        let mut specifier = first_arg
-            .map(|arg| self.emit_ir_fragment_to_string(&self.convert_expression(arg)))
-            .unwrap_or_default();
-        let mut prefix = String::new();
-
-        if first_arg.is_some() && !first_arg_is_string_like {
-            let temp = self.generate_hoisted_temp();
-            prefix = format!("{temp} = {specifier}, ");
-            specifier = temp;
-        }
-
-        if matches!(self.module_kind, ModuleKind::System) {
-            return IRNode::Raw(format!("context_1.import({specifier})").into());
-        }
-
-        let amd_branch = self.dynamic_import_amd_branch(&specifier);
-        if matches!(self.module_kind, ModuleKind::UMD) {
-            return IRNode::Raw(
-                format!(
-                    "{prefix}__syncRequire ? {} : {amd_branch}",
-                    self.dynamic_import_commonjs_branch(&specifier)
-                )
-                .into(),
-            );
-        }
-
-        IRNode::Raw(format!("{prefix}{amd_branch}").into())
-    }
-
-    fn first_dynamic_import_argument(&self, args: Option<&NodeList>) -> Option<NodeIndex> {
-        args?
-            .nodes
-            .iter()
-            .copied()
-            .find(|&idx| crate::transforms::emit_utils::call_argument_should_emit(self.arena, idx))
-    }
-
-    fn emit_ir_fragment_to_string(&self, ir: &IRNode) -> String {
-        let mut printer = if let Some(source_text) = self.source_text {
-            IRPrinter::with_arena_and_source(self.arena, source_text)
-        } else {
-            IRPrinter::with_arena(self.arena)
-        };
-        if let Some(transforms) = self.transforms.as_ref() {
-            printer.set_transforms(transforms.clone());
-        }
-        printer.emit(ir).to_string()
-    }
-
-    fn dynamic_import_commonjs_branch(&self, specifier: &str) -> String {
-        crate::transforms::emit_utils::dynamic_import_cjs_form(specifier)
-    }
-
-    fn dynamic_import_amd_branch(&self, specifier: &str) -> String {
-        let id = self.dynamic_import_promise_counter.get();
-        self.dynamic_import_promise_counter.set(id + 1);
-        format!(
-            "new Promise(function (resolve_{id}, reject_{id}) {{ require([{specifier}], resolve_{id}, reject_{id}); }}).then(__importStar)"
-        )
-    }
-
-    /// Check if a call expression callee is super.method or super[expr] and transform to
-    /// _super.prototype.method.call(this, args) or _super.prototype[expr].call(this, args)
-    fn try_convert_super_method_call(
-        &self,
-        callee_idx: NodeIndex,
-        args: Vec<IRNode>,
-        is_optional_call: bool,
-    ) -> Option<IRNode> {
-        let callee_node = self.arena.get(callee_idx)?;
-
-        // Check for super.method(args) → _super.prototype.method.call(this, args)
-        // In static context: super.method(args) → _super.method.call(this, args)
-        if callee_node.kind == syntax_kind_ext::PROPERTY_ACCESS_EXPRESSION {
-            let access = self.arena.get_access_expr(callee_node)?;
-            let obj_node = self.arena.get(access.expression)?;
-            if obj_node.kind == SyntaxKind::SuperKeyword as u16 {
-                let method_name = get_identifier_text(self.arena, access.name_or_argument)?;
-                let super_proto_method = if self.is_static.get() {
-                    // Static: _super.method
-                    IRNode::PropertyAccess {
-                        object: Box::new(IRNode::id(self.super_name.clone())),
-                        property: method_name.into(),
-                    }
-                } else {
-                    // Instance: _super.prototype.method
-                    IRNode::PropertyAccess {
-                        object: Box::new(IRNode::PropertyAccess {
-                            object: Box::new(IRNode::id(self.super_name.clone())),
-                            property: "prototype".to_string().into(),
-                        }),
-                        property: method_name.into(),
-                    }
-                };
-                if is_optional_call {
-                    return Some(self.convert_optional_super_method_call(super_proto_method, args));
-                }
-                let call_method = IRNode::PropertyAccess {
-                    object: Box::new(super_proto_method),
-                    property: "call".to_string().into(),
-                };
-                let mut call_args = vec![self.current_this_ir()];
-                call_args.extend(args);
-                return Some(IRNode::CallExpr {
-                    callee: Box::new(call_method),
-                    arguments: call_args,
-                });
-            }
-        }
-
-        // Check for super[expr](args) → _super.prototype[expr].call(this, args)
-        // In static context: super[expr](args) → _super[expr].call(this, args)
-        if callee_node.kind == syntax_kind_ext::ELEMENT_ACCESS_EXPRESSION {
-            let access = self.arena.get_access_expr(callee_node)?;
-            let obj_node = self.arena.get(access.expression)?;
-            if obj_node.kind == SyntaxKind::SuperKeyword as u16 {
-                let index_expr = self.convert_expression(access.name_or_argument);
-                let super_base = if self.is_static.get() {
-                    IRNode::id(self.super_name.clone())
-                } else {
-                    IRNode::PropertyAccess {
-                        object: Box::new(IRNode::id(self.super_name.clone())),
-                        property: "prototype".to_string().into(),
-                    }
-                };
-                let super_proto_elem = IRNode::ElementAccess {
-                    object: Box::new(super_base),
-                    index: Box::new(index_expr),
-                };
-                if is_optional_call {
-                    return Some(self.convert_optional_super_method_call(super_proto_elem, args));
-                }
-                let call_method = IRNode::PropertyAccess {
-                    object: Box::new(super_proto_elem),
-                    property: "call".to_string().into(),
-                };
-                let mut call_args = vec![self.current_this_ir()];
-                call_args.extend(args);
-                return Some(IRNode::CallExpr {
-                    callee: Box::new(call_method),
-                    arguments: call_args,
-                });
-            }
-        }
-
-        None
-    }
-
-    fn convert_optional_super_method_call(&self, receiver: IRNode, args: Vec<IRNode>) -> IRNode {
-        let temp = self.generate_hoisted_temp();
-        let temp_ref = || IRNode::id(temp.clone());
-
-        let mut call_args = vec![self.current_this_ir()];
-        call_args.extend(args);
-
-        IRNode::ConditionalExpr {
-            condition: Box::new(IRNode::logical_or(
-                IRNode::binary(
-                    IRNode::assign(temp_ref(), receiver).paren(),
-                    "===",
-                    IRNode::NullLiteral,
-                ),
-                IRNode::binary(temp_ref(), "===", IRNode::Undefined),
-            )),
-            when_true: Box::new(IRNode::Undefined),
-            when_false: Box::new(IRNode::CallExpr {
-                callee: Box::new(IRNode::PropertyAccess {
-                    object: Box::new(temp_ref()),
-                    property: "call".to_string().into(),
-                }),
-                arguments: call_args,
-            }),
-        }
-    }
-
-    fn convert_new_expression(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        // NewExpression uses CallExprData (same as CallExpression)
-        if let Some(call_data) = self.arena.get_call_expr(node) {
-            let callee = self.convert_expression(call_data.expression);
-            let args = if let Some(ref args) = call_data.arguments {
-                args.nodes
-                    .iter()
-                    .map(|&a| self.convert_expression(a))
-                    .collect()
-            } else {
-                vec![]
-            };
-            IRNode::NewExpr {
-                callee: Box::new(callee),
-                arguments: args,
-                explicit_arguments: call_data.arguments.is_some(),
-            }
-        } else {
-            IRNode::ASTRef(idx)
-        }
-    }
-
-    fn convert_property_access(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        // PropertyAccessExpression uses AccessExprData
-        if let Some(access) = self.arena.get_access_expr(node) {
-            // Check for super.property → _super.prototype.property (instance) or _super.property (static)
-            if let Some(obj_node) = self.arena.get(access.expression)
-                && obj_node.kind == SyntaxKind::SuperKeyword as u16
-                && let Some(name) = get_identifier_text(self.arena, access.name_or_argument)
-            {
-                return if self.is_static.get() {
-                    IRNode::PropertyAccess {
-                        object: Box::new(IRNode::id(self.super_name.clone())),
-                        property: name.into(),
-                    }
-                } else {
-                    IRNode::PropertyAccess {
-                        object: Box::new(IRNode::PropertyAccess {
-                            object: Box::new(IRNode::id(self.super_name.clone())),
-                            property: "prototype".to_string().into(),
-                        }),
-                        property: name.into(),
-                    }
-                };
-            }
-
-            let object = self.convert_expression(access.expression);
-            if let Some(name) = get_identifier_text(self.arena, access.name_or_argument) {
-                return IRNode::PropertyAccess {
-                    object: Box::new(object),
-                    property: name.into(),
-                };
-            }
-        }
-        IRNode::ASTRef(idx)
-    }
-
-    fn convert_element_access(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        // ElementAccessExpression uses AccessExprData
-        if let Some(access) = self.arena.get_access_expr(node) {
-            // Check for super[expr] → _super.prototype[expr] (instance) or _super[expr] (static)
-            if let Some(obj_node) = self.arena.get(access.expression)
-                && obj_node.kind == SyntaxKind::SuperKeyword as u16
-            {
-                let index = self.convert_expression(access.name_or_argument);
-                let super_base = if self.is_static.get() {
-                    IRNode::id(self.super_name.clone())
-                } else {
-                    IRNode::PropertyAccess {
-                        object: Box::new(IRNode::id(self.super_name.clone())),
-                        property: "prototype".to_string().into(),
-                    }
-                };
-                return IRNode::ElementAccess {
-                    object: Box::new(super_base),
-                    index: Box::new(index),
-                };
-            }
-
-            let object = self.convert_expression(access.expression);
-            let index = self.convert_expression(access.name_or_argument);
-            IRNode::ElementAccess {
-                object: Box::new(object),
-                index: Box::new(index),
-            }
-        } else {
-            IRNode::ASTRef(idx)
-        }
-    }
-
-    fn convert_binary_expression(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        if let Some(bin) = self.arena.get_binary_expr(node) {
-            let left = self.convert_expression(bin.left);
-            let right = self.convert_expression(bin.right);
-            let op = self.get_binary_operator(bin.operator_token);
-
-            // Handle logical operators specially
-            if op == "||" {
-                return IRNode::LogicalOr {
-                    left: Box::new(left),
-                    right: Box::new(right),
-                };
-            }
-            if op == "&&" {
-                return IRNode::LogicalAnd {
-                    left: Box::new(left),
-                    right: Box::new(right),
-                };
-            }
-
-            IRNode::BinaryExpr {
-                left: Box::new(left),
-                operator: op.into(),
-                right: Box::new(right),
-            }
-        } else {
-            IRNode::ASTRef(idx)
-        }
-    }
-
-    fn get_binary_operator(&self, token: u16) -> String {
-        crate::transforms::emit_utils::operator_to_str(token).to_string()
-    }
-
-    fn convert_prefix_unary(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        // PrefixUnaryExpression uses UnaryExprData
-        if let Some(unary) = self.arena.get_unary_expr(node) {
-            let operand = self.convert_expression(unary.operand);
-            let op = self.get_prefix_operator(unary.operator);
-            IRNode::PrefixUnaryExpr {
-                operator: op.into(),
-                operand: Box::new(operand),
-            }
-        } else {
-            IRNode::ASTRef(idx)
-        }
-    }
-
-    fn get_prefix_operator(&self, token: u16) -> String {
-        crate::transforms::emit_utils::operator_to_str(token).to_string()
-    }
-
-    fn convert_postfix_unary(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        // PostfixUnaryExpression uses UnaryExprData
-        if let Some(unary) = self.arena.get_unary_expr(node) {
-            let operand = self.convert_expression(unary.operand);
-            let op = match unary.operator {
-                k if k == SyntaxKind::PlusPlusToken as u16 => "++".to_string(),
-                k if k == SyntaxKind::MinusMinusToken as u16 => "--".to_string(),
-                _ => "".to_string(),
-            };
-            IRNode::PostfixUnaryExpr {
-                operand: Box::new(operand),
-                operator: op.into(),
-            }
-        } else {
-            IRNode::ASTRef(idx)
-        }
-    }
-
-    fn convert_parenthesized(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        if let Some(paren) = self.arena.get_parenthesized(node) {
-            IRNode::Parenthesized(Box::new(self.convert_expression(paren.expression)))
-        } else {
-            IRNode::ASTRef(idx)
-        }
-    }
-
-    fn convert_conditional(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        // ConditionalExpression uses ConditionalExprData
-        if let Some(cond) = self.arena.get_conditional_expr(node) {
-            IRNode::ConditionalExpr {
-                condition: Box::new(self.convert_expression(cond.condition)),
-                when_true: Box::new(self.convert_expression(cond.when_true)),
-                when_false: Box::new(self.convert_expression(cond.when_false)),
-            }
-        } else {
-            IRNode::ASTRef(idx)
-        }
     }
 
     fn convert_array_literal(&self, idx: NodeIndex) -> IRNode {
@@ -2938,48 +1865,5 @@ impl<'a> AstToIr<'a> {
             left.kind == syntax_kind_ext::OBJECT_LITERAL_EXPRESSION
                 || left.kind == syntax_kind_ext::ARRAY_LITERAL_EXPRESSION
         })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::AstToIr;
-    use tsz_parser::parser::syntax_kind_ext;
-
-    #[test]
-    fn attached_line_comment_start_allows_statement_trailing_comments() {
-        assert_eq!(
-            AstToIr::attached_line_comment_start(syntax_kind_ext::VARIABLE_STATEMENT, " // ok"),
-            Some(1)
-        );
-        assert_eq!(
-            AstToIr::attached_line_comment_start(syntax_kind_ext::VARIABLE_STATEMENT, "; // ok"),
-            Some(2)
-        );
-    }
-
-    #[test]
-    fn attached_line_comment_start_rejects_comments_after_parent_delimiters() {
-        assert_eq!(
-            AstToIr::attached_line_comment_start(
-                syntax_kind_ext::VARIABLE_STATEMENT,
-                " } // not inner",
-            ),
-            None
-        );
-        assert_eq!(
-            AstToIr::attached_line_comment_start(
-                syntax_kind_ext::VARIABLE_STATEMENT,
-                " }) // not inner",
-            ),
-            None
-        );
-    }
-
-    #[test]
-    fn can_attach_trailing_comment_gap_rejects_parent_delimiters() {
-        assert!(AstToIr::can_attach_trailing_comment_gap(" ; "));
-        assert!(!AstToIr::can_attach_trailing_comment_gap(" } "));
-        assert!(!AstToIr::can_attach_trailing_comment_gap(" }) "));
     }
 }
